+++ IndoorIOS/CADMeViewController.h
@@ -26,7 +26,7 @@
 @property (nonatomic, strong) NSMutableData *jsonData;
 @property (nonatomic, strong) NSOperationQueue *queue;
 
-@property (nonatomic) int maxTimeUnitCount;
+@property (nonatomic) NSInteger maxTimeUnitCount;
 
 // the set of IconDownloader objects for each image
 @property (nonatomic, strong) NSMutableDictionary *imageDownloadsInProgress;
