+++ IndoorIOS/Util/RSADataSigner.m
@@ -29,7 +29,7 @@
 
 - (NSString *)formatPrivateKey:(NSString *)privateKey {
     const char *pstr = [privateKey UTF8String];
-    int len = [privateKey length];
+    NSInteger len = [privateKey length];
     NSMutableString *result = [NSMutableString string];
     [result appendString:@"-----BEGIN PRIVATE KEY-----\n"];
     int index = 0;
@@ -71,10 +71,10 @@
 	[formatKey writeToFile:path atomically:YES encoding:NSUTF8StringEncoding error:nil];
 	
 	const char *message = [string cStringUsingEncoding:NSUTF8StringEncoding];
-    int messageLength = strlen(message);
+    NSInteger messageLength = strlen(message);
     unsigned char *sig = (unsigned char *)malloc(256);
-	unsigned int sig_len;
-    int ret = rsa_sign_with_private_key_pem((char *)message, messageLength, sig, &sig_len, (char *)[path UTF8String]);
+	NSUInteger sig_len;
+    NSInteger ret = rsa_sign_with_private_key_pem((char *)message, messageLength, sig, &sig_len, (char *)[path UTF8String]);
 	//签名成功,需要给签名字符串base64编码和UrlEncode,该两个方法也可以根据情况替换为自己函数
     if (ret == 1) {
         NSString * base64String = base64StringFromData([NSData dataWithBytes:sig length:sig_len]);
