+++ IndoorIOS/Util/openssl_wrapper.m
@@ -17,8 +17,8 @@
 #include <string.h>
 
 
-int rsa_sign_with_private_key_pem(char *message, int message_length
-                                  , unsigned char *signature, unsigned int *signature_length
+NSInteger rsa_sign_with_private_key_pem(char *message, NSInteger message_length
+                                  , unsigned char *signature, NSUInteger *signature_length
                                   , char *private_key_file_path)
 {
     unsigned char sha1[20];
@@ -50,8 +50,8 @@ int rsa_sign_with_private_key_pem(char *message, int message_length
     return success;
 }
 
-int rsa_verify_with_public_key_pem(char *message, int message_length
-                                   , unsigned char *signature, unsigned int signature_length
+NSInteger rsa_verify_with_public_key_pem(char *message, NSInteger message_length
+                                   , unsigned char *signature, NSUInteger signature_length
                                    , char *public_key_file_path)
 {
     unsigned char sha1[20];
@@ -76,7 +76,7 @@ int rsa_verify_with_public_key_pem(char *message, int message_length
 
 NSString *base64StringFromData(NSData *signature)
 {
-    int signatureLength = [signature length];
+    NSInteger signatureLength = [signature length];
     unsigned char *outputBuffer = (unsigned char *)malloc(2 * 4 * (signatureLength / 3 + 1));
     int outputLength = EVP_EncodeBlock(outputBuffer, [signature bytes], signatureLength);
     outputBuffer[outputLength] = '\0';
@@ -87,13 +87,13 @@ NSString *base64StringFromData(NSData *signature)
 
 NSData *dataFromBase64String(NSString *base64String)
 {
-    int stringLength = [base64String length];
+    NSInteger stringLength = [base64String length];
     const unsigned char *strBuffer = (const unsigned char *)[base64String UTF8String];
     unsigned char *outputBuffer = (unsigned char *)malloc(2 * 3 * (stringLength / 4 + 1));
     int outputLength = EVP_DecodeBlock(outputBuffer, strBuffer, stringLength);
     
     int zeroByteCounter = 0;
-    for (int i = stringLength - 1; i >= 0; i--)
+    for (NSInteger i = stringLength - 1; i >= 0; i--)
     {
         if (strBuffer[i] == '=')
         {
@@ -113,11 +113,11 @@ NSData *dataFromBase64String(NSString *base64String)
 NSString *rsaSignString(NSString *stringToSign, NSString *privateKeyFilePath, BOOL *signSuccess)
 {
     const char *message = [stringToSign cStringUsingEncoding:NSUTF8StringEncoding];
-    int messageLength = strlen(message);
+    NSInteger messageLength = strlen(message);
     unsigned char *sig = (unsigned char *)malloc(256);
-    unsigned int sig_len;
+    NSUInteger sig_len;
     char *filePath = (char *)[privateKeyFilePath cStringUsingEncoding:NSUTF8StringEncoding];
-    int sign_ok = rsa_sign_with_private_key_pem((char *)message, messageLength
+    NSInteger sign_ok = rsa_sign_with_private_key_pem((char *)message, messageLength
                                                 , sig, &sig_len
                                                 , filePath);
     NSString *signedString = nil;
@@ -137,12 +137,12 @@ NSString *rsaSignString(NSString *stringToSign, NSString *privateKeyFilePath, BO
 void rsaVerifyString(NSString *stringToVerify, NSString *signature, NSString *publicKeyFilePath, BOOL *verifySuccess)
 {
     const char *message = [stringToVerify cStringUsingEncoding:NSUTF8StringEncoding];
-    int messageLength = [stringToVerify lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
+    NSInteger messageLength = [stringToVerify lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
     NSData *signatureData = dataFromBase64String(signature);
     unsigned char *sig = (unsigned char *)[signatureData bytes];
-    unsigned int sig_len = [signatureData length];
+    NSUInteger sig_len = [signatureData length];
     char *filePath = (char *)[publicKeyFilePath cStringUsingEncoding:NSUTF8StringEncoding];
-    int verify_ok = rsa_verify_with_public_key_pem((char *)message, messageLength
+    NSInteger verify_ok = rsa_verify_with_public_key_pem((char *)message, messageLength
                                                    , sig, sig_len
                                                    , filePath);
     if (1 == verify_ok)
@@ -159,7 +159,7 @@ NSString *formattedPEMString(NSString *originalString)
 {    
     NSString *trimmedString = [originalString stringByReplacingOccurrencesOfString:@"\n" withString:@""];
     const char *c = [trimmedString UTF8String];
-    int len = [trimmedString length];
+    NSInteger len = [trimmedString length];
     NSMutableString *result = [NSMutableString string];
     [result appendString:@"-----BEGIN PRIVATE KEY-----\n"];
     int index = 0;
